backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
httpx==0.28.1
openai==2.31.0
python-dotenv==1.2.2